GetShardMapping bug fix Separate Main method for Demo purpose ToString methods for Shard and Shard Map Initial methods to try and get both List and Range Shards Initial methods to create Shard Map Manger and Shards

+++ azure-samples/src/main/java/com/microsoft/azure/elasticdb/samples/elasticscalestarterkit/SqlDatabaseUtils.java
@@ -10,6 +10,7 @@ import java.sql.DriverManager;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
+import java.util.concurrent.TimeUnit;
 
 /**
  * Helper methods for interacting with SQL Databases.
@@ -24,13 +25,14 @@ public final class SqlDatabaseUtils {
      * Returns true if we can connect to the database.
      */
     public static boolean TryConnectToSqlDatabase() {
-        String connectionString = Configuration.GetConnectionString(Configuration.getShardMapManagerServerName(), MasterDatabaseName);
+        String serverName = Configuration.getShardMapManagerServerName();
+        String connectionString = Configuration.GetConnectionString(serverName, MasterDatabaseName);
 
         SQLServerConnection conn = null;
         try {
-            ConsoleUtils.WriteInfo("Connecting to Azure Portal...\r\n");
+            ConsoleUtils.WriteInfo("Connecting to Azure Portal...");
             conn = (SQLServerConnection) DriverManager.getConnection(connectionString);
-            ConsoleUtils.WriteInfo("Connection Successful...\r\n");
+            ConsoleUtils.WriteInfo("Connection Successful... Server Name: " + serverName);
         } catch (Exception e) {
             ConsoleUtils.WriteWarning("Failed to connect to SQL database with connection string:");
             System.out.printf("\n%1$s\n" + "\r\n", connectionString);
@@ -54,7 +56,8 @@ public final class SqlDatabaseUtils {
         }
     }
 
-    public static boolean DatabaseExists(String connectionString, String dbName) {
+    public static boolean DatabaseExists(String serverName, String dbName) {
+        String connectionString = Configuration.GetConnectionString(serverName, dbName);
         SQLServerConnection conn = null;
         try {
             conn = (SQLServerConnection) DriverManager.getConnection(connectionString);
@@ -76,7 +79,46 @@ public final class SqlDatabaseUtils {
         return true;
     }
 
-    public static void CreateDatabase(String shardMapManagerServerName, String databaseName) {
+    public static void CreateDatabase(String server, String db) {
+        ConsoleUtils.WriteInfo("Creating database %s", db);
+        SQLServerConnection conn = null;
+        String connectionString = Configuration.GetConnectionString(server, MasterDatabaseName);
+        try {
+            conn = (SQLServerConnection) DriverManager.getConnection(connectionString);
+            String query = "SELECT CAST(SERVERPROPERTY('EngineEdition') AS NVARCHAR(128))";
+            try (Statement stmt = conn.createStatement()) {
+                ResultSet rs = stmt.executeQuery(query);
+                if (rs.next() && rs.getInt(1) == 5) {
+                    query = String.format("CREATE DATABASE %1$s (EDITION = '%2$s')",
+                            BracketEscapeName(db), Configuration.getDatabaseEdition());
+                    stmt.executeUpdate(query);
+                    while (!DatabaseIsOnline(conn, BracketEscapeName(db))) {
+                        ConsoleUtils.WriteInfo("Waiting for database %s to come online...", db);
+                        TimeUnit.SECONDS.sleep(5);
+                    }
+                    ConsoleUtils.WriteInfo("Database %s is online", db);
+                }
+            } catch (SQLException ex) {
+                ex.printStackTrace();
+            }
+        } catch (Exception e) {
+            ConsoleUtils.WriteWarning("Failed to connect to SQL database with connection string:");
+            System.out.printf("\n%1$s\n" + "\r\n", connectionString);
+            ConsoleUtils.WriteWarning("If this connection string is incorrect, please update the Configuration file.\n\nException message: %s", e.getMessage());
+        } finally {
+            connFinally(conn);
+        }
+    }
+
+    public static boolean DatabaseIsOnline(SQLServerConnection conn, String db) {
+        try (Statement stmt = conn.createStatement()) {
+            ResultSet rs = stmt.executeQuery("SELECT COUNT(*) FROM sys.databases WHERE name = '"
+                    + db + "' and state = 0");
+            return (rs.next() && rs.getInt(1) > 0);
+        } catch (SQLException ex) {
+            ex.printStackTrace();
+            return false;
+        }
     }
 
     public static void ExecuteSqlScript(String shardMapManagerServerName, String databaseName, String initializeShardScriptFile) {
@@ -95,7 +137,7 @@ public final class SqlDatabaseUtils {
             String query = "SELECT CAST(SERVERPROPERTY('EngineEdition') AS NVARCHAR(128))";
             try (Statement stmt = conn.createStatement()) {
                 ResultSet rs = stmt.executeQuery(query);
-                while (rs.next()) {
+                if (rs.next()) {
                     query = rs.getInt(1) == 5 ?
                             String.format("DROP DATABASE %1$s", BracketEscapeName(db))
                             : String.format("ALTER DATABASE %1$s SET SINGLE_USER WITH ROLLBACK IMMEDIATE"
